+++ utils.js
@@ -1,6 +1,7 @@
 // Dependencies
 const { promises: { access, }, readdirSync, writeFileSync, } = require('fs');
 const { parse, } = require('json2csv');
+const csvtojsonV2 = require('csvtojson');
 
 /**
  * Check if the given file exists
@@ -76,7 +77,36 @@ const reviewJSONToCsv = (jsonInput) => {
     }
 };
 
+/**
+ * Extract the name, url, and id of the resto from a csv file
+ * @param {String} csvFilePath - The location of the csv file 
+ * @returns {Promise<Object | Error>} - The parsed json object or error message
+ */
+const restoCsvToJSON = async (csvFilePath) => {
+    try {
 
-module.exports = { fileExists, combine, reviewJSONToCsv, };
+        // Read the csv file
+        const parsedJson = await csvtojsonV2().fromFile(csvFilePath);
 
-console.log(reviewJSONToCsv(combine()));
+        // Extract the fields
+        const processed = parsedJson.map(resto => {
+            return {
+                name: resto.name,
+                webUrl: resto.webUrl,
+                id: resto.id,
+            };
+        });
+
+        // Write to JSON file
+        writeFileSync(`${csvFilePath}.json`, JSON.stringify(processed, null, 2));
+
+        return processed;
+
+    } catch (err) {
+        throw err;
+    }
+
+};
+
+
+module.exports = { fileExists, combine, reviewJSONToCsv, restoCsvToJSON, };
